Add CA profile for OpenVPN root CA and regular CA's
This commit introduces the concepts of profiles in certificate management. The idea is that we expose a set of prefilled fields which we consider safe for basic use. Each profile is tailored for a specific use case. Right now this commit introduces a profile for OpenVPN root CA exposing basic cert extensions and some other fields which the user can use to build his payload for creating a root CA. The UI can also make use of this by allowing user to select a profile and filling those fields beforehand so the user does not has to know the finer details of everything and if he does, then that means his use case is complex and he can change the fields to as he wishes.